# public/openapi.yml
openapi: 3.0.3
info:
  title: Control Ventas API (BFF)
  version: 0.2.0
  description: Contrato de endpoints internos Next.js (mock en tests, Supabase en runtime).
servers:
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - DemoRole: []
...
paths:
...
  /api/dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Obtener resumen del dashboard
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Supabase lee `daily_sales_summary`, `low_stock_products` y ventas pendientes.
      x-permission: dashboard.view
      responses:
        "200":
          $ref: "#/components/responses/DashboardSummaryResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/dashboard/sales-trend:
    get:
      tags: [Dashboard]
      summary: Obtener tendencia de ventas para el grafico del dashboard
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Agrega ventas por dia en el rango `from`/`to`.
      x-permission: dashboard.view
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          $ref: "#/components/responses/ReportObjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/dashboard/recent-sales:
    get:
      tags: [Dashboard]
      summary: Obtener ultimas ventas
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Pagina con `skip`/`limit`.
      x-permission: dashboard.view
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/SaleArrayResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/dashboard/low-stock:
    get:
      tags: [Dashboard]
      summary: Obtener stock bajo compacto
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Supabase lee la vista `low_stock_products`.
      x-permission: dashboard.view
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          $ref: "#/components/responses/ReportArrayResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/categories:
    get:
      tags: [Categories]
      summary: Listar categorias
      description: Usa mock o Supabase segun `API_DATA_SOURCE`.
      x-permission: products.view
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Lista paginada de categorias
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/PaginatedList"
                      - type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Categories]
      summary: Crear categoria
      description: Usa mock o Supabase segun `API_DATA_SOURCE`.
      x-permission: products.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Categoria creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Categories]
      summary: Obtener categoria
      x-permission: products.view
      responses:
        "200":
          description: Categoria
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Categories]
      summary: Actualizar categoria
      description: Usa mock o Supabase segun `API_DATA_SOURCE`.
      x-permission: products.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdateInput"
      responses:
        "200":
          description: Categoria actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Categories]
      summary: Borrar categoria (soft delete)
      description: Mock devuelve `deleted=true`. Supabase marca `is_active=false`.
      x-permission: products.manage
      responses:
        "200":
          description: Categoria borrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/products:
    get:
      tags: [Products]
      summary: Listar productos
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Pagina con `skip`/`limit` y `count` exacto.
      x-permission: products.view
      parameters:
        - $ref: "#/components/parameters/Skip"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filtra productos activos o inactivos.
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
          description: Filtra productos por categoria.
      responses:
        "200":
          description: Lista paginada de productos
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/PaginatedList"
                      - type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/Product"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Products]
      summary: Crear producto
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. SKU duplicado responde `409`.
      x-permission: products.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateInput"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/products/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Products]
      summary: Obtener producto
      x-permission: products.view
      responses:
        "200":
          description: Producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Products]
      summary: Actualizar producto
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. SKU duplicado responde `409`.
      x-permission: products.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateInput"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Products]
      summary: Borrar producto (soft delete)
      description: Mock devuelve `deleted=true`. Supabase marca `is_active=false`.
      x-permission: products.manage
      responses:
        "200":
          description: Producto borrado
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Product"
                      - type: object
                        properties:
                          deleted:
                            type: boolean
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/products/{id}/price:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Products]
      summary: Cambiar precio de producto
      description: Mock simula historial. Supabase usa RPC `update_product_price`.
      x-permission: products.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPriceInput"
      responses:
        "200":
          $ref: "#/components/responses/ReportObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/products/{id}/price-history:
    parameters:
      - $ref: "#/components/parameters/Id"
      - $ref: "#/components/parameters/Skip"
      - $ref: "#/components/parameters/Limit"
    get:
      tags: [Products]
      summary: Obtener historial de precios
      description: Usa mock o Supabase segun `API_DATA_SOURCE`. Pagina con `skip`/`limit`.
      x-permission: products.view
      responses:
        "200":
          description: Historial paginado de precios
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/PaginatedList"
                      - type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/ProductPriceHistory"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/products/{id}/suppliers:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Products]
      summary: Listar proveedores de un producto
      x-permission: products.view
      responses:
        "200":
          $ref: "#/components/responses/SupplierProductArrayResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/inventory:
    get:
      tags: [Inventory]
      summary: Listar stock actual
      x-permission: inventory.view
      parameters:
        - name: lowStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Productos en inventario
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
  /api/inventory/movements:
    get:
      tags: [Inventory]
      summary: Listar movimientos de inventario
      x-permission: inventory.view
      parameters:
        - name: productId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Movimientos de stock
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StockMovement"
  /api/inventory/stock-card:
    get:
      tags: [Inventory]
      summary: Obtener kardex de inventario
      x-permission: inventory.view
      parameters:
        - name: productId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Movimientos del kardex
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StockMovement"
  /api/inventory/adjustments:
    post:
      tags: [Inventory]
      summary: Crear ajuste de inventario simulado
      x-permission: inventory.manage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustmentInput"
      responses:
        "201":
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovementResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/sales:
    get:
      tags: [Sales]
      summary: Listar ventas
      x-permission: sales.view
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SaleStatus"
        - name: customerId
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de ventas
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sale"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Sales]
      summary: Crear venta simulada
      x-permission: sales.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleCreateInput"
      responses:
        "201":
          description: Venta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/sales/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Sales]
      summary: Obtener venta
      x-permission: sales.view
      responses:
        "200":
          description: Venta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/sales/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
    patch:
      tags: [Sales]
      summary: Cancelar venta mock
      x-permission: sales.create
      responses:
        "200":
          $ref: "#/components/responses/ReportObjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/sales/{id}/return:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Sales]
      summary: Registrar devolucion de cliente
      description: Devuelve una venta mediante RPC `return_sale` cuando `API_DATA_SOURCE=supabase`.
      x-permission: sales.create
      responses:
        "200":
          $ref: "#/components/responses/ReportObjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/sales/{id}/receipt:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Sales]
      summary: Obtener recibo o factura
      description: Payload compuesto de venta, items y cliente para impresion o vista de recibo.
      x-permission: sales.view
      responses:
        "200":
          $ref: "#/components/responses/ReportObjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/purchases/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Purchases]
      summary: Obtener compra
      x-permission: purchases.view
      responses:
        "200":
          description: Compra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/reports/daily-sales:
    get:
      tags: [Reports]
      summary: Reporte diario de ventas
      x-permission: reports.view
      responses:
        "200":
          $ref: "#/components/responses/ReportArrayResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    DemoRole:
      type: apiKey
      in: header
      name: x-demo-role
      description: "Rol temporal para endpoints mock: admin, vendedor, almacen o contador."
  parameters:
...
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    Skip:
      name: skip
      in: query
      required: false
      description: Registros a omitir desde el inicio del resultado filtrado.
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      required: false
      description: Cantidad por pagina. Minimo 10, default 10.
      schema:
        type: integer
        minimum: 10
        default: 10
  responses:
...
    Forbidden:
      description: Sin permiso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    DashboardSummaryResponse:
      description: Resumen del dashboard
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                $ref: "#/components/schemas/DashboardSummary"
...
    ReportObjectResponse:
      description: Objeto generico
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                type: object
...